# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  isAdmin?: boolean;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  /** @nullable */
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface Project {
  id: number;
  slug: string;
  titre: string;
  categorie: string;
  statut: string;
  annee: number;
  duree: string;
  langue: string;
  synopsis: string;
  intention: string;
  image: string;
  sortOrder: number;
  featured?: boolean;
}

export interface CreateProjectBody {
  slug: string;
  titre: string;
  categorie: string;
  statut: string;
  annee: number;
  duree: string;
  langue: string;
  synopsis: string;
  intention: string;
  image: string;
  sortOrder?: number;
  featured?: boolean;
}

export interface UpdateProjectBody {
  slug?: string;
  titre?: string;
  categorie?: string;
  statut?: string;
  annee?: number;
  duree?: string;
  langue?: string;
  synopsis?: string;
  intention?: string;
  image?: string;
  sortOrder?: number;
  featured?: boolean;
}

export interface NewsArticle {
  id: number;
  titre: string;
  categorie: string;
  dateLabel: string;
  excerpt: string;
  image: string;
  sortOrder: number;
}

export interface CreateNewsBody {
  titre: string;
  categorie: string;
  dateLabel: string;
  excerpt: string;
  image: string;
  sortOrder?: number;
}

export interface UpdateNewsBody {
  titre?: string;
  categorie?: string;
  dateLabel?: string;
  excerpt?: string;
  image?: string;
  sortOrder?: number;
}

export interface Service {
  id: number;
  titre: string;
  description: string;
  icon: string;
  sortOrder: number;
}

export interface CreateServiceBody {
  titre: string;
  description: string;
  icon: string;
  sortOrder?: number;
}

export interface UpdateServiceBody {
  titre?: string;
  description?: string;
  icon?: string;
  sortOrder?: number;
}

export interface Partner {
  id: number;
  nom: string;
  description: string;
  sortOrder: number;
}

export interface CreatePartnerBody {
  nom: string;
  description: string;
  sortOrder?: number;
}

export interface UpdatePartnerBody {
  nom?: string;
  description?: string;
  sortOrder?: number;
}

export interface SiteSettings {
  id: number;
  heroBadge: string;
  heroTitleLine1: string;
  heroTitleLine2: string;
  heroSubtitle: string;
  quoteText: string;
  quoteAuthor: string;
  statsProjets: string;
  statsPays: string;
  statsFestivals: string;
  aboutHistoire: string;
  aboutVision: string;
  founderName: string;
  founderTitle: string;
  founderBio: string;
  founderImage: string;
  contactEmail: string;
  contactPhone: string;
  contactAddress: string;
}

export interface UpdateSiteSettingsBody {
  heroBadge?: string;
  heroTitleLine1?: string;
  heroTitleLine2?: string;
  heroSubtitle?: string;
  quoteText?: string;
  quoteAuthor?: string;
  statsProjets?: string;
  statsPays?: string;
  statsFestivals?: string;
  aboutHistoire?: string;
  aboutVision?: string;
  founderName?: string;
  founderTitle?: string;
  founderBio?: string;
  founderImage?: string;
  contactEmail?: string;
  contactPhone?: string;
  contactAddress?: string;
}