# lib/api-zod/src/generated/types/createNewsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNewsBody {
  titre: string;
  categorie: string;
  dateLabel: string;
  excerpt: string;
  image: string;
  sortOrder?: number;
}