# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  slug?: string;
  titre?: string;
  categorie?: string;
  statut?: string;
  annee?: number;
  duree?: string;
  langue?: string;
  synopsis?: string;
  intention?: string;
  image?: string;
  sortOrder?: number;
  featured?: boolean;
}